% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lasso.R
\name{boot.lasso}
\alias{boot.lasso}
\title{Bootstrap for LASSO.}
\usage{
boot.lasso(lambdas, x, y, beta0 = rep(0, ncol(x)), ts = 0.1,
  method = lasso_cd, k = 5, n.sim = 100)
}
\arguments{
\item{lambdas}{values of lambda}

\item{x}{Predictors}

\item{y}{Response}

\item{beta0}{Initial guess of beta}

\item{ts}{Stepsize}

\item{method}{Optimization method}

\item{k}{No. of folds}

\item{n.sim}{No. of bootstrap replicates}
}
\description{
Bootstrap for LASSO.
}
