% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{cv.lasso}
\alias{cv.lasso}
\title{Cross-validation Tools
cross-validation}
\usage{
cv.lasso(lambdas, x, y, wt = NULL, ts = NULL, method = lasso_cd,
  k = 5, n_it = 10, df = NULL)
}
\arguments{
\item{lambdas}{penalty parameter}

\item{x}{predictors}

\item{y}{response}

\item{wt}{weights for the coefficients of weighted LASSO.}

\item{ts}{sequence of stepsize (for lasso_sg, lasso_pg)}

\item{method}{lasso optimization function}

\item{k}{number of fold}

\item{n_it}{number of iteration for lasso_cd method}

\item{df}{Degree of freedom. Number of desired variables to be zero.}

\item{beta0}{initial guess of beta}
}
\description{
Cross-validation Tools
cross-validation
}
