% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lasso.R
\name{example.boot}
\alias{example.boot}
\title{Example}
\usage{
example.boot(n.sim = 200)
}
\arguments{
\item{n.sim}{No. of bootstrap replicates. Default value is 200.}
}
\description{
Example to check boot-strap LASSO
}
