% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{example.cv}
\alias{example.cv}
\title{Examples}
\usage{
example.cv(rel_er)
}
\arguments{
\item{rel_er}{Relative error. Should be between 0 (minimum error) and 1 (maximum error).}
}
\description{
Examples
}
