% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_lasso.R
\name{example.sensitivity}
\alias{example.sensitivity}
\title{Example}
\usage{
example.sensitivity(wt = NULL, nsim = 200)
}
\arguments{
\item{wt}{weights for the coefficients of weighted LASSO. Defaults to NULL}

\item{nsim}{number simulations with perturbed weights. Default value is 200.}
}
\description{
Example to check cross-validation for LASSO.
}
