% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_lasso.R
\name{example_lasso_1}
\alias{example_lasso_1}
\title{Example}
\usage{
example_lasso_1(wt = NULL)
}
\arguments{
\item{wt}{weights for the coefficients of weighted LASSO. Defaults to NULL}
}
\description{
Example 1 of LASSO using different optimization techniques
}
