% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_lasso.R
\name{glmnet_optim_plot}
\alias{glmnet_optim_plot}
\title{glmnet plot}
\usage{
glmnet_optim_plot(lambdas, x, y)
}
\arguments{
\item{lambdas}{Values of the penalty term}

\item{x}{Predictors}

\item{y}{Response}
}
\description{
glmnet plot
}
