% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprecise_lasso.R
\name{imp.lasso}
\alias{imp.lasso}
\title{Lower and upper estimates of LASSO coefficients}
\usage{
imp.lasso(lambda, x, y, wtl, wtu, ts = NULL, method = lasso_optim_cd,
  n_it = 10)
}
\arguments{
\item{lambda}{value of the penalty parameter}

\item{x}{predictors}

\item{y}{response}

\item{wtl}{lower bound of the weights for the coefficients of weighted LASSO}

\item{wtu}{upper bound of the weights for the coefficients of weighted LASSO}

\item{ts}{stepsize for proximal gradient and sub-gradient method (use opt_ts() to generate stepsize). Defaults to NULL}

\item{method}{lasso optimization function.  Three different methods are available to use. method = c(lasso_optim_cd, lasso_optim_sg, lasso_optim_pg). Defaults to lasso_optim_cd}

\item{n_it}{number of iteration for lasso_optim_cd method. Default value is 10.}
}
\value{
The function returns the lower and upper estimates of LASSO coefficients.
}
\description{
Lower and upper estimates of LASSO coefficients using optimization over weights
}
