% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{lasso.plot}
\alias{lasso.plot}
\title{lasso plots}
\usage{
lasso.plot(cv, main = NULL)
}
\arguments{
\item{cv}{model obtained by doing cross-validation.}

\item{main}{Title of the plot.}
}
\description{
lasso plots
}
