% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_lasso.R
\name{lasso_cd_plot}
\alias{lasso_cd_plot}
\title{cd plot}
\usage{
lasso_cd_plot(lambdas, x, y, n_it = 100, wt)
}
\arguments{
\item{lambdas}{Values of the penalty term}

\item{x}{Predictors}

\item{y}{Response}

\item{n_it}{Number of iterations. Default value 100}

\item{wt}{weights for the coefficients of weighted LASSO.}

\item{beta0}{Initial guess of the regression coefficients}
}
\description{
Plots the co-efficient path using co-ordinate descent optimization
}
