% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{normalize}
\alias{normalize}
\title{Normalizing the data}
\usage{
normalize(x, y)
}
\arguments{
\item{x}{predictor}

\item{y}{response}
}
\value{
The function returns following list of outputs
\item{x}{normalized predictor}
\item{y}{normalized response}
}
\description{
Function for normalizing the dataset
}
