% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lasso.R
\name{pg_optim}
\alias{pg_optim}
\title{Proximal gradient optimization.}
\usage{
pg_optim(x, f, df, pg, ts)
}
\arguments{
\item{x}{Starting value.}

\item{f}{Function to optimize.}

\item{df}{Gradient of f.}

\item{pg}{Proximal operator for g.}

\item{ts}{Sequence of step sizes. use opt_ts() for stepsize generation}
}
\value{
The function returns the optimal solution
}
\description{
Function for proximal-gradient optimization for non-differentiable functions
We're trying to minimize f + g; where f is the differentiable part and g is the non-differentiable part.
}
