% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_lasso.R
\name{sensitivity_lasso}
\alias{sensitivity_lasso}
\title{Perturbation of weights}
\usage{
sensitivity_lasso(lambdas, x, y, wt = NULL, ts = NULL,
  method = lasso_cd, k = 5, n_it = 10, df = NULL, nsim = 50)
}
\arguments{
\item{lambdas}{values of the penalty parameter}

\item{x}{predictors}

\item{y}{response}

\item{wt}{weights for the coefficients of weighted LASSO. Defaults to NULL}

\item{ts}{stepsize for proximal gradient and sub-gradient method (use opt_ts() to generate stepsize). Defaults to NULL}

\item{method}{lasso optimization function.  Three different methods are available to use. method = c(lasso_cd, lasso_sg, lasso_pg). Defaults to lasso_cd}

\item{k}{number of fold. Default value is 5.}

\item{n_it}{number of iteration for lasso_cd method. Default value is 10.}

\item{df}{Degree of freedom. Number of desired variables to be zero. Defaults to NULL}

\item{nsim}{number of perturbed weights. Default value is 50}
}
\value{
The function returns the summary of sensitivity analysis.
}
\description{
Sensitivity analysis of predictors using perturbed weights
}
