% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lasso.R
\name{sg_optim}
\alias{sg_optim}
\title{Sub-gradient optimization.}
\usage{
sg_optim(x, f, df, ts)
}
\arguments{
\item{x}{Starting value.}

\item{f}{Function to optimize.}

\item{df}{Any subgradient of f.}

\item{ts}{Sequence of step sizes. use opt_ts() for stepsize generation}
}
\value{
The function returns the optimal solution
}
\description{
Function for sub-gradient optimization for non-differentiable functions
}
