% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lasso.R
\name{boot.lasso}
\alias{boot.lasso}
\title{Bootstrap for LASSO.}
\usage{
boot.lasso(lambdas, x, y, wt = NULL, ts = NULL, method = lasso_cd,
  k = 5, n_it = 10, df = NULL, n.sim = 500)
}
\arguments{
\item{lambdas}{values of the penalty parameter}

\item{x}{predictors}

\item{y}{response}

\item{wt}{weights for the coefficients of weighted LASSO. Defaults to NULL}

\item{ts}{stepsize for proximal gradient and sub-gradient method (use opt_ts() to generate stepsize). Defaults to NULL}

\item{method}{optimization method. Three different methods are available to use. method = c(lasso_cd, lasso_sg, lasso_pg). Defaults to lasso_cd}

\item{k}{no. of folds for cross-validation. Default value is 5.}

\item{n_it}{no. of iteration for lasso_cd optimization. Default value is 10.}

\item{df}{degree of freedom. Number of desired variables to be zero. Defaults to NULL}

\item{n.sim}{no. of bootstrap replicates. Default value is 500.}
}
\value{
The summary frame of the bootstrap comprising mean, median, bias, standard deviation and confidence intervals.
}
\description{
Function for getting bootstrap estimates for LASSO.
}
