% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lasso.R
\name{cd_optim}
\alias{cd_optim}
\title{Co-ordinate descent optimization.}
\usage{
cd_optim(x, f, v, s, n_it)
}
\arguments{
\item{x}{Starting value.}

\item{f}{Function to optimize.}

\item{n_it}{Number of iterations. Default value 100}

\item{df}{Any subgradient of f.}
}
\description{
Co-ordinate descent optimization.
}
