% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lasso.R
\name{cd_optim}
\alias{cd_optim}
\title{Co-ordinate descent optimization.}
\usage{
cd_optim(x, f, x_it, n_it = 100)
}
\arguments{
\item{x}{Starting value.}

\item{f}{Function to optimize.}

\item{x_it}{Iterative x obtained from the subgradient function}

\item{n_it}{Number of iterations. Default value is 100}
}
\value{
The function returns the optimal solution
}
\description{
Function for co-ordinate descent optimization for non-differentiable functions
}
