% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{cv.lasso}
\alias{cv.lasso}
\title{Cross-validation Tools
cross-validation}
\usage{
cv.lasso(lambda, x, y, beta0, ts, method, k = 5, n_it = 100)
}
\arguments{
\item{lambda}{penalty parameter}

\item{x}{predictors}

\item{y}{response}

\item{beta0}{initial guess of beta}

\item{ts}{sequence of stepsize (for lasso_sg, lasso_pg)}

\item{method}{lasso optimization function}

\item{k}{number of fold}

\item{n_it}{number of iteration for lasso_cd method}
}
\description{
Cross-validation Tools
cross-validation
}
