% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{cv.mse}
\alias{cv.mse}
\title{root mean-squared error}
\usage{
cv.mse(original, estimate)
}
\arguments{
\item{original}{original value}

\item{estimate}{estimated value}
}
\description{
root mean-squared error
}
