% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{cv.plot}
\alias{cv.plot}
\title{cv curve plot}
\usage{
cv.plot(cv, main = NULL)
}
\arguments{
\item{cv}{model obtained by doing cross-validation.}

\item{main}{Title of the plot.}
}
\description{
cv curve plot
}
