% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{cv.random.partition}
\alias{cv.random.partition}
\title{Random partition}
\usage{
cv.random.partition(x, y, k)
}
\arguments{
\item{x}{predictors}

\item{y}{response}

\item{k}{number of fold}
}
\value{
Partitioned dataset
}
\description{
For internal use
}
