% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{lasso.plot}
\alias{lasso.plot}
\title{lasso plots}
\usage{
lasso.plot(output, index, main = NULL)
}
\arguments{
\item{output}{model obtained by doing cross-validation.}

\item{index}{index obtained from cross-validation for checking desired lambdas intersect.}

\item{main}{Title of the plot.}
}
\description{
lasso plots
}
