% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_lasso.R
\name{lasso_cd}
\alias{lasso_cd}
\title{Coefficient path (cd)}
\usage{
lasso_cd(lambdas, x, y, beta0, n_it = 100, ...)
}
\arguments{
\item{lambdas}{Values of the penalty term}

\item{x}{Predictors}

\item{y}{Response}

\item{beta0}{Initial guess of the regression coefficients}

\item{n_it}{Number of iterations. Default value 100.}
}
\description{
Coefficient path (cd)
}
