% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_lasso.R
\name{lasso_cd}
\alias{lasso_cd}
\title{Coefficient path (cd)}
\usage{
lasso_cd(lambdas, x, y, n_it = 100, wt, ...)
}
\arguments{
\item{lambdas}{Values of the penalty term}

\item{x}{Predictors}

\item{y}{Response}

\item{n_it}{Number of iterations. Default value 100.}

\item{wt}{weights for the coefficients of weighted LASSO.}
}
\value{
The function returns the coefficient of the predictors for each lambda.
}
\description{
This function is to be used for getting co-efficient path using co-ordinate descent optimization
}
