% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_lasso.R
\name{lasso_cd_plot}
\alias{lasso_cd_plot}
\title{cd plot}
\usage{
lasso_cd_plot(x, y, n_it = 100, wt)
}
\arguments{
\item{x}{Predictors}

\item{y}{Response}

\item{n_it}{Number of iterations. Default value 100}

\item{wt}{weights for the coefficients of weighted LASSO.}
}
\description{
cd plot
}
