% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_lasso.R
\name{lasso_optim_cd}
\alias{lasso_optim_cd}
\title{LASSO optimization using co-ordinate descent}
\usage{
lasso_optim_cd(lambda, x, y, n_it = 100, wt)
}
\arguments{
\item{lambda}{Penalty term}

\item{x}{Predictors}

\item{y}{Response}

\item{n_it}{Number of iterations. Default value 100}

\item{wt}{weights for the coefficients of weighted LASSO.}
}
\value{
The function returns the coefficient of the predictors.
}
\description{
This function is to be used for solving LASSO optimization using co-ordinate descent method
}
