% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_lasso.R
\name{lasso_pg_plot}
\alias{lasso_pg_plot}
\title{pg plot}
\usage{
lasso_pg_plot(lambdas, x, y, beta0, ts, wt)
}
\arguments{
\item{lambdas}{Values of the penalty term}

\item{x}{Predictors}

\item{y}{Response}

\item{beta0}{Initial guess of the regression coefficients}

\item{ts}{Stepsize for optimization method}

\item{wt}{weights for the coefficients of weighted LASSO.}
}
\description{
Plots the co-efficient path using proximal-gradient optimization
}
