% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lasso.R
\name{opt_ts}
\alias{opt_ts}
\title{Sequence of step sizes for optimization.}
\usage{
opt_ts(t, m, n)
}
\arguments{
\item{t}{Starting value.}

\item{m}{Number of constant steps.}

\item{n}{Number of diminishing steps.}
}
\value{
The sequence of stepsize
}
\description{
Function to generate stepsize for sub-gradient optimization and proximal-gradient optimization
}
