% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lasso.R
\name{pg_optim}
\alias{pg_optim}
\title{Proximal gradient optimization.}
\usage{
pg_optim(x, f, df, g, p, ts)
}
\arguments{
\item{x}{Starting value.}

\item{f}{Function to optimize.}

\item{df}{Gradient of f.}

\item{p}{Proximal operator for g.}

\item{ts}{Sequence of step sizes.}
}
\description{
We're trying to minimize f + g; prox is the proximal operator for g.
}
