% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lasso.R
\name{sg_optim}
\alias{sg_optim}
\title{Subgradient optimization.}
\usage{
sg_optim(x, f, df, ts)
}
\arguments{
\item{x}{Starting value.}

\item{f}{Function to optimize.}

\item{df}{Any subgradient of f.}

\item{ts}{Sequence of step sizes.}
}
\description{
Subgradient optimization.
}
